revert me: change adjacency matrix dtype

+++ stellargraph/core/graph.py
@@ -787,15 +787,11 @@ class StellarGraph:
         tgt_idx = index.to_iloc(self._edges.targets[selector])
         n = len(index)
 
-        adj = sps.csr_matrix(
-            (weights, (src_idx, tgt_idx)), shape=(n, n), dtype="float32"
-        )
+        adj = sps.csr_matrix((weights, (src_idx, tgt_idx)), shape=(n, n))
         if not self.is_directed():
             # in an undirected graph, the adjacency matrix should be symmetric: which means counting
             # weights from either "incoming" or "outgoing" edges, but not double-counting self loops
-            backward = sps.csr_matrix(
-                (weights, (tgt_idx, src_idx)), shape=(n, n), dtype="float32"
-            )
+            backward = sps.csr_matrix((weights, (tgt_idx, src_idx)), shape=(n, n))
             backward.setdiag(0)
             adj += backward
